Add files via upload
I was working on two different ideas.  The first was a fairly limited project, with little or no possible future applications.  This one - which i hope will be my submission - includes an additional function, which (if it works) will allow future players to be added, and all their relevant info will be imported to the dataframe